Split Quiz Show, Bunny Hop, and Rap Men game engines into their own C files.
