+++ src/engines/bunny_hop.c
@@ -1,0 +1,109 @@
+#include "src/engines/bunny_hop.h"
+
+asm(".include \"include/gba.inc\""); // Temporary
+
+
+/* BUNNY HOP */
+
+
+#include "asm/scenes/bunny_hop/asm_08037f84.s"
+
+#include "asm/scenes/bunny_hop/asm_08037f94.s"
+
+#include "asm/scenes/bunny_hop/asm_08037fd4.s"
+
+#include "asm/scenes/bunny_hop/asm_08038000.s"
+
+#include "asm/scenes/bunny_hop/asm_08038244.s"
+
+#include "asm/scenes/bunny_hop/asm_08038248.s"
+
+#include "asm/scenes/bunny_hop/asm_080382ac.s"
+
+#include "asm/scenes/bunny_hop/asm_080382b0.s"
+
+#include "asm/scenes/bunny_hop/asm_080382f4.s"
+
+#include "asm/scenes/bunny_hop/asm_08038314.s"
+
+#include "asm/scenes/bunny_hop/asm_0803833c.s"
+
+#include "asm/scenes/bunny_hop/asm_08038350.s"
+
+#include "asm/scenes/bunny_hop/asm_08038384.s"
+
+#include "asm/scenes/bunny_hop/asm_080383f0.s"
+
+#include "asm/scenes/bunny_hop/asm_08038414.s"
+
+#include "asm/scenes/bunny_hop/asm_08038418.s"
+
+#include "asm/scenes/bunny_hop/asm_0803841c.s"
+
+#include "asm/scenes/bunny_hop/asm_08038438.s"
+
+#include "asm/scenes/bunny_hop/asm_0803843c.s"
+
+#include "asm/scenes/bunny_hop/asm_080384b8.s"
+
+#include "asm/scenes/bunny_hop/asm_08038598.s"
+
+#include "asm/scenes/bunny_hop/asm_080385f4.s"
+
+#include "asm/scenes/bunny_hop/asm_080386e0.s"
+
+#include "asm/scenes/bunny_hop/asm_080387c4.s"
+
+#include "asm/scenes/bunny_hop/asm_080388a8.s"
+
+#include "asm/scenes/bunny_hop/asm_080388ac.s"
+
+#include "asm/scenes/bunny_hop/asm_080388b0.s"
+
+#include "asm/scenes/bunny_hop/asm_080388b4.s"
+
+#include "asm/scenes/bunny_hop/asm_080388d8.s"
+
+#include "asm/scenes/bunny_hop/asm_0803899c.s"
+
+#include "asm/scenes/bunny_hop/asm_08038a84.s"
+
+#include "asm/scenes/bunny_hop/asm_08038b98.s"
+
+#include "asm/scenes/bunny_hop/asm_08038ce0.s"
+
+#include "asm/scenes/bunny_hop/asm_08038d18.s"
+
+#include "asm/scenes/bunny_hop/asm_08038d54.s"
+
+#include "asm/scenes/bunny_hop/asm_08038ef8.s"
+
+#include "asm/scenes/bunny_hop/asm_08038f2c.s"
+
+#include "asm/scenes/bunny_hop/asm_08038fbc.s"
+
+#include "asm/scenes/bunny_hop/asm_08039128.s"
+
+#include "asm/scenes/bunny_hop/asm_08039164.s"
+
+#include "asm/scenes/bunny_hop/asm_0803934c.s"
+
+#include "asm/scenes/bunny_hop/asm_08039388.s"
+
+#include "asm/scenes/bunny_hop/asm_08039404.s"
+
+#include "asm/scenes/bunny_hop/asm_08039440.s"
+
+#include "asm/scenes/bunny_hop/asm_080394a4.s"
+
+#include "asm/scenes/bunny_hop/asm_080395dc.s"
+
+#include "asm/scenes/bunny_hop/asm_0803960c.s"
+
+#include "asm/scenes/bunny_hop/asm_08039698.s"
+
+#include "asm/scenes/bunny_hop/asm_08039738.s"
+
+#include "asm/scenes/bunny_hop/asm_0803978c.s"
+
+#include "asm/scenes/bunny_hop/asm_080397f8.s"
