+++ src/engines/bunny_hop.h
@@ -1,0 +1,69 @@
+#pragma once
+
+#include "global.h"
+#include "sound.h"
+
+// Engine Macros/Enums:
+
+// OAM Animations:
+
+// Palettes:
+
+// Sound Effects:
+
+// Engine Data:
+
+// Engine Definition Data:
+
+// Functions:
+// extern ? func_08037f84(?);
+// extern ? func_08037f94(?);
+// extern ? func_08037fd4(?);
+// extern ? func_08038000(?);
+// extern ? func_08038244(?);
+// extern ? func_08038248(?);
+// extern ? func_080382ac(?);
+// extern ? func_080382b0(?);
+// extern ? func_080382f4(?);
+// extern ? func_08038314(?);
+// extern ? func_0803833c(?);
+// extern ? func_08038350(?);
+// extern ? func_08038384(?);
+// extern ? func_080383f0(?);
+// extern ? func_08038414(?);
+// extern ? func_08038418(?);
+// extern ? func_0803841c(?);
+// extern ? func_08038438(?);
+// extern ? func_0803843c(?);
+// extern ? func_080384b8(?);
+// extern ? func_08038598(?);
+// extern ? func_080385f4(?);
+// extern ? func_080386e0(?);
+// extern ? func_080387c4(?);
+// extern ? func_080388a8(?);
+// extern ? func_080388ac(?);
+// extern ? func_080388b0(?);
+// extern ? func_080388b4(?);
+// extern ? func_080388d8(?);
+// extern ? func_0803899c(?);
+// extern ? func_08038a84(?);
+// extern ? func_08038b98(?);
+// extern ? func_08038ce0(?);
+// extern ? func_08038d18(?);
+// extern ? func_08038d54(?);
+// extern ? func_08038ef8(?);
+// extern ? func_08038f2c(?);
+// extern ? func_08038fbc(?);
+// extern ? func_08039128(?);
+// extern ? func_08039164(?);
+// extern ? func_0803934c(?);
+// extern ? func_08039388(?);
+// extern ? func_08039404(?);
+// extern ? func_08039440(?);
+// extern ? func_080394a4(?);
+// extern ? func_080395dc(?);
+// extern ? func_0803960c(?);
+// extern ? func_08039698(?);
+// extern ? func_08039738(?);
+// extern ? func_0803978c(?);
+// extern ? func_080397f8(?);
