+++ src/engines/quiz_show.c
@@ -1,0 +1,89 @@
+#include "src/engines/quiz_show.h"
+
+asm(".include \"include/gba.inc\""); // Temporary
+
+
+/* QUIZ SHOW */
+
+
+#include "asm/scenes/quiz_show/asm_0803709c.s"
+
+#include "asm/scenes/quiz_show/asm_08037178.s"
+
+#include "asm/scenes/quiz_show/asm_08037280.s"
+
+#include "asm/scenes/quiz_show/asm_080372c0.s"
+
+#include "asm/scenes/quiz_show/asm_080372e8.s"
+
+#include "asm/scenes/quiz_show/asm_08037378.s"
+
+#include "asm/scenes/quiz_show/asm_080373ac.s"
+
+#include "asm/scenes/quiz_show/asm_080373dc.s"
+
+#include "asm/scenes/quiz_show/asm_080373f4.s"
+
+#include "asm/scenes/quiz_show/asm_08037404.s"
+
+#include "asm/scenes/quiz_show/asm_08037434.s"
+
+#include "asm/scenes/quiz_show/asm_08037460.s"
+
+#include "asm/scenes/quiz_show/asm_08037858.s"
+
+#include "asm/scenes/quiz_show/asm_0803785c.s"
+
+#include "asm/scenes/quiz_show/asm_080378d8.s"
+
+#include "asm/scenes/quiz_show/asm_08037a64.s"
+
+#include "asm/scenes/quiz_show/asm_08037a78.s"
+
+#include "asm/scenes/quiz_show/asm_08037be0.s"
+
+#include "asm/scenes/quiz_show/asm_08037bf4.s"
+
+#include "asm/scenes/quiz_show/asm_08037cb8.s"
+
+#include "asm/scenes/quiz_show/asm_08037cd0.s"
+
+#include "asm/scenes/quiz_show/asm_08037cf4.s"
+
+#include "asm/scenes/quiz_show/asm_08037d00.s"
+
+#include "asm/scenes/quiz_show/asm_08037d0c.s"
+
+#include "asm/scenes/quiz_show/asm_08037d1c.s"
+
+#include "asm/scenes/quiz_show/asm_08037d2c.s"
+
+#include "asm/scenes/quiz_show/asm_08037d6c.s"
+
+#include "asm/scenes/quiz_show/asm_08037d90.s"
+
+#include "asm/scenes/quiz_show/asm_08037dc8.s"
+
+#include "asm/scenes/quiz_show/asm_08037e24.s"
+
+#include "asm/scenes/quiz_show/asm_08037e4c.s"
+
+#include "asm/scenes/quiz_show/asm_08037e50.s"
+
+#include "asm/scenes/quiz_show/asm_08037e54.s"
+
+#include "asm/scenes/quiz_show/asm_08037e70.s"
+
+#include "asm/scenes/quiz_show/asm_08037e74.s"
+
+#include "asm/scenes/quiz_show/asm_08037e78.s"
+
+#include "asm/scenes/quiz_show/asm_08037e7c.s"
+
+#include "asm/scenes/quiz_show/asm_08037e80.s"
+
+#include "asm/scenes/quiz_show/asm_08037f58.s"
+
+#include "asm/scenes/quiz_show/asm_08037f5c.s"
+
+#include "asm/scenes/quiz_show/asm_08037f60.s"
