+++ src/engines/quiz_show.h
@@ -1,0 +1,59 @@
+#pragma once
+
+#include "global.h"
+#include "sound.h"
+
+// Engine Macros/Enums:
+
+// OAM Animations:
+
+// Palettes:
+
+// Sound Effects:
+
+// Engine Data:
+
+// Engine Definition Data:
+
+// Functions:
+// extern ? func_0803709c(?);
+// extern ? func_08037178(?);
+// extern ? func_08037280(?);
+// extern ? func_080372c0(?);
+// extern ? func_080372e8(?);
+// extern ? func_08037378(?);
+// extern ? func_080373ac(?);
+// extern ? func_080373dc(?);
+// extern ? func_080373f4(?);
+// extern ? func_08037404(?);
+// extern ? func_08037434(?);
+// extern ? func_08037460(?);
+// extern ? func_08037858(?);
+// extern ? func_0803785c(?);
+// extern ? func_080378d8(?);
+// extern ? func_08037a64(?);
+// extern ? func_08037a78(?);
+// extern ? func_08037be0(?);
+// extern ? func_08037bf4(?);
+// extern ? func_08037cb8(?);
+// extern ? func_08037cd0(?);
+// extern ? func_08037cf4(?);
+// extern ? func_08037d00(?);
+// extern ? func_08037d0c(?);
+// extern ? func_08037d1c(?);
+// extern ? func_08037d2c(?);
+// extern ? func_08037d6c(?);
+// extern ? func_08037d90(?);
+// extern ? func_08037dc8(?);
+// extern ? func_08037e24(?);
+// extern ? func_08037e4c(?);
+// extern ? func_08037e50(?);
+// extern ? func_08037e54(?);
+// extern ? func_08037e70(?);
+// extern ? func_08037e74(?);
+// extern ? func_08037e78(?);
+// extern ? func_08037e7c(?);
+// extern ? func_08037e80(?);
+// extern ? func_08037f58(?);
+// extern ? func_08037f5c(?);
+// extern ? func_08037f60(?);
