+++ src/scenes.h
@@ -3,128 +3,6 @@
 #include "global.h"
 #include "sound.h"
 
-// Quiz Show
-
-// extern ? func_0803709c(?);
-// extern ? func_08037178(?);
-// extern ? func_08037280(?);
-// extern ? func_080372c0(?);
-// extern ? func_080372e8(?);
-// extern ? func_08037378(?);
-// extern ? func_080373ac(?);
-// extern ? func_080373dc(?);
-// extern ? func_080373f4(?);
-// extern ? func_08037404(?);
-// extern ? func_08037434(?);
-// extern ? func_08037460(?);
-// extern ? func_08037858(?);
-// extern ? func_0803785c(?);
-// extern ? func_080378d8(?);
-// extern ? func_08037a64(?);
-// extern ? func_08037a78(?);
-// extern ? func_08037be0(?);
-// extern ? func_08037bf4(?);
-// extern ? func_08037cb8(?);
-// extern ? func_08037cd0(?);
-// extern ? func_08037cf4(?);
-// extern ? func_08037d00(?);
-// extern ? func_08037d0c(?);
-// extern ? func_08037d1c(?);
-// extern ? func_08037d2c(?);
-// extern ? func_08037d6c(?);
-// extern ? func_08037d90(?);
-// extern ? func_08037dc8(?);
-// extern ? func_08037e24(?);
-// extern ? func_08037e4c(?);
-// extern ? func_08037e50(?);
-// extern ? func_08037e54(?);
-// extern ? func_08037e70(?);
-// extern ? func_08037e74(?);
-// extern ? func_08037e78(?);
-// extern ? func_08037e7c(?);
-// extern ? func_08037e80(?);
-// extern ? func_08037f58(?);
-// extern ? func_08037f5c(?);
-// extern ? func_08037f60(?);
-
-// Bunny Hop
-
-// extern ? func_08037f84(?);
-// extern ? func_08037f94(?);
-// extern ? func_08037fd4(?);
-// extern ? func_08038000(?);
-// extern ? func_08038244(?);
-// extern ? func_08038248(?);
-// extern ? func_080382ac(?);
-// extern ? func_080382b0(?);
-// extern ? func_080382f4(?);
-// extern ? func_08038314(?);
-// extern ? func_0803833c(?);
-// extern ? func_08038350(?);
-// extern ? func_08038384(?);
-// extern ? func_080383f0(?);
-// extern ? func_08038414(?);
-// extern ? func_08038418(?);
-// extern ? func_0803841c(?);
-// extern ? func_08038438(?);
-// extern ? func_0803843c(?);
-// extern ? func_080384b8(?);
-// extern ? func_08038598(?);
-// extern ? func_080385f4(?);
-// extern ? func_080386e0(?);
-// extern ? func_080387c4(?);
-// extern ? func_080388a8(?);
-// extern ? func_080388ac(?);
-// extern ? func_080388b0(?);
-// extern ? func_080388b4(?);
-// extern ? func_080388d8(?);
-// extern ? func_0803899c(?);
-// extern ? func_08038a84(?);
-// extern ? func_08038b98(?);
-// extern ? func_08038ce0(?);
-// extern ? func_08038d18(?);
-// extern ? func_08038d54(?);
-// extern ? func_08038ef8(?);
-// extern ? func_08038f2c(?);
-// extern ? func_08038fbc(?);
-// extern ? func_08039128(?);
-// extern ? func_08039164(?);
-// extern ? func_0803934c(?);
-// extern ? func_08039388(?);
-// extern ? func_08039404(?);
-// extern ? func_08039440(?);
-// extern ? func_080394a4(?);
-// extern ? func_080395dc(?);
-// extern ? func_0803960c(?);
-// extern ? func_08039698(?);
-// extern ? func_08039738(?);
-// extern ? func_0803978c(?);
-// extern ? func_080397f8(?);
-
-// Rap Men
-
-extern u32 *func_080398b4(u32);
-extern void func_080398d4(void);
-extern void func_080398e4(void);
-extern void func_08039924(void);
-extern void func_08039950(u32);
-extern void func_08039a40(void);
-extern void func_08039a40(void);
-extern void func_08039a98(u32);
-extern void func_08039a98(u32 arg0);
-extern void func_08039ad4(void);
-extern void func_08039aa4(void);
-extern void func_08039ad8(u32, struct struct_080179f4_sub *, u32);
-extern u32 func_08039b2c(u32, u32, u32);
-extern void func_08039b48(void);
-extern void func_08039b4c(u32, struct struct_080179f4_sub *);
-extern void func_08039c00(void);
-extern void func_08039c60(void);
-extern void func_08039cb8(void);
-extern void func_08039d10(void);
-extern void func_08039d7c(char *);
-extern void func_08039df8(void);
-
 // Rat Race
 
 // extern ? func_08039dfc(?);
